Move base.py for table to __init__
